Ask for 2mb stack, not 2mb heap
I guess when we were copying this from Jagex's config we typoed Xss, which changes the maximum stack size, to Xms which changes minimum heap size. Apparently the GC tries to keep it's heap as close to minimum as possible when Xms is specified, so it was running almost constantly on release.