Remove faulty hw acceleration check
The hardware acceleration check was always false unless someone had
global JVM arguments modified (or was modyfing launcher JVM arguments)
because the check was actually checking if hardware acceleration is
already enabled and not if it is supported.
To solve this, default hardware acceleration mode to OFF, but leave the
possibility to easily change it to OPENGL or DIRECTDRAW to tune
performance.

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>