Support Dark Mode for the native window chrome on macOS
macOS 10.14 added a dark mode for the native window chrome, but applications need to declare that they support it for it to be used. This can be done through `NSRequiresAquaSystemAppearance` in the `Info.plist` file. Whether the light or dark theme gets used is automatically decided based on the user's settings.