osx: fix freezing on launch
MacOS adds a "process serial number" argument when the launcher is launched from Finder the first time after download, and in some cases (such as on Big Sur beta 1) also on subsequent launches. This causes the option parser in the launcher to error, resulting in the launcher app freezing in the system dock.

This commit registers the -psn argument with the option parser so that the parser will no longer throw a UnrecognizedOptionException when the argument is added.