snap: Bundle GVFS to fix web links not opening
The vanilla client fails to open web browsers on Linux systems in the event GVFS isn't installed. These are not covered by the xdg-open fallback functionality that RuneLite's plugins use.
This commit is enough to ensure every snap user has enough of GFVS to catch the cases the xdg-open functionality doesn't cover.