snap: Bundle notify-send, arm64 support, remove pulseaudio interface
notify-send has been explicitly added to the stage-package to enable desktop notifications.

arm64 will probably work assuming the build process is run on an arm64 host (real or emulated).
Such a snap file could be submitted to the store where it would be delivered to arm64 clients as expected.
This hasn't actually been tested to work however.

The Pulseaudio interface has been removed. It was added in the previous snap patch to fix legacy
snapd clients that didn't understand the "audio-playback" interface. However, the Gnome-3-38
extension since enforces a minimum snapd version that ensures all users wouldn't need
the legacy "pulseaudio" interface. The functional difference is that audio-playback
does not grant access to the microphone unlike Pulseaudio.