Add scale argument
RuneLite uses Windows scaling by default, which is not typical of Java
since Java's manifest claims dpi awareness, and awt will attempt to
scale itself to the system dpi factory. RuneLite's manifest does not
include this and that is enough to prevent it from being able to detect
the system scale factor.

This argument allows using the awt scaling introduced in Java 9 by
having the process mark itself as dpi aware and manually setting the j2d
scaling factor. It is required to mark the process as dpi aware to
prevent Windows from trying to scale the application even after awt has
done its scaling.