Set java2d props prior to swing startup
The --mode argument has been broken since 2.0 with nojvm due to it
applying the prop after splash screen init, which is too late since
swing startup is when the props are read.

This also moves the other vm props to be set prior to splash screen
startup for simplicity, but it isn't required.

The non-prop vm args have been split into their own list, which is
currently just the crash file path, used for writing the packr vmargs
and for the jvm launcher.