Fix parsing macos psn
joptsimple by default allows abbreviations for long options if the
abbreviation is not ambiguous. The passed psn was parsed as an
abbreviation to the "psn" argument since they are passed in short option
form (one -). With the addition of the "postinstall" argument the psn
argument becomes ambiguous, preventing it from being parsed, and is
being passed to the client causing the client to error.

Disable the joptsimple abbreviations feature, and accept the "p"
option on macos for psn