Remove java.net.preferIPv4Stack
THis is set too late currently. Both libnet and PlainSocketImpl both
read this prop, but libnet reads it OnLoad which is during VM
initialization, so we end up running with libnet having this set to
false while PlainSocketImpl has it set to true.

This ends up causing InetAddress to pick the v6 networking stack,
resolving addresses to ipv6 ips, and then PlainSocketImpl erroring when
it encounters an non-ipv4 address when in ipv4 only mode.

Since this seemingly has never really worked just remove it.