add forking launcher
This makes the launcher relaunch itself with the desired vm arguments
for the client, instead of relying on the launcher vm args to be the
same as the client args.

runelite.launcher.nojvm was renamed to runelite.launcher.reflect to
avoid the launcher using reflect mode from updating the packr vmargs
required for the previous launchers. The only launcher still using this
is the appimage launcher, which requires it due to the appimage
unmounting the filesystem after it exits.