use both Windows root trust store and bundled trust store
Instead of having to pick one trust store or the other, use both
truststores. On some older systems like Windows 7, the cloudflare root
is not trusted by the system, causing the launcher to not work without
telling it to use the jre truststore.

This makes the --use-jre-truststore option no longer necessary.