innosetup: fix invalid drive error with missing installs
When the innosetup app path in the registry points to an invalid
location, like a drive which no longer exists, setup breaks because it
is unable to install to the location, but also the uninstaller no longer
exists too and so it can't be uninstalled. This forces the user to fix
it via editing the registry.

This could be worked around with DisableDirPage=no, which always forces
the dir page to be shown. But this would allow users to install multiple
times to different places, which is not desirable.

Reset the dir path if it is pointed to a non-default location which
doesn't exist. Note the dir page is still not shown in this case, but at
least it will fix the install enough to let the user uninstall it again
if they want to move it to a different location.