filesystem permissions: limit number of startup permissions checks
This is way too slow with a large number of files. The majority of the
files on my system are screenshots, which are not important since we do
not read or write from them.

Most of the time seems to be in isWritable() which is doing GetFileSecurity(),
OpenThreadToken(), and AccessCheck().

Limit the checks to the main top level directories and then also limit
the number of files checked in each directory.