improve fatal error dialog net error window
The new http client seems to wrap exceptions in IOException, so some of
these previous checks were no longer working. Traverse the exception
stack backwards as the most useful error messages are always at the end.

Add a check for the certificate being not valid, which is usually
indicative of the clock being off.